phr-manager/requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-JWT-Extended==4.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.3
requests==2.25.1
urllib3==1.26.4
Werkzeug==1.0.1
